sysif: define passwd symbol for group stuff.
* sysif.c (sysif_init): The passwd symbol is used by both the
passwd and group structure, so we need to initialize it if
either HAVE_PWUID or HAVE_GRGID is set. Certain Android SDK
levels have getgrgid but not getpwuid, so it's possible to end
up with HAVE_GRGID but not HAVE_PWUID, in which case passwd_s
ends up nil, blowing up the make_struct_type call.