fix crash if built-in variable is unbound.
We use lookup_var_l in many places to look up the current
dynamic value of a built-in variable such as *stdout*.
Those places assume that a a valid location is returned
which can be subject to a deref. If the application calls
makunbound to remove such a variable, that deref will
crash due to a null pointer dereference.

Possible repro steps are numerous, possible for many variables.
One example:

  (makunbound '*stdout*)
  (put-line)


* eval.c (lookukp_var_l): If the binding is not found, do
not return a nulloc, but throw an error exception.