bug: symbol-value place always global.
We have a problem. If v is a dynamic variable, then
the form

  (let (v)
    (set (symbol-value 'v) 3))

is not behaving correctly; it's updating the top-level
value of v not the rebound one.

* eval.c (set_symbol_value): New static function.
(eval_init): Register sys:set-symbol-value intrinsic.
The top-vb variable, though no longer referenced by
the symbol-value place, because existing compiled
code depends on it.

* stdlib/place.tl (symbol-value): Rewrite the place
logic to use symbol-value to access the variable,
and set-symbol-value to update it, instead of referencing
sys:top-vb.
(sys:get-vb): This function has to stay, because it
provides run-time support for code compiled with the
buggy version of the place.

* tests/019/symbol-value.tl: New file.