load: now establishes a block named load.
Files loaded from the command line or via the load function
or @(load) directive now have a block named load in scope.
Thus (return-from load <expr>) may be used to abort a
load when it is finished. The load function will then
return the value of <expr>.

* eval.c (load): Bind a load block around the whole thing
and use the captured return value as the function's
return value.

* match.c (v_load): Bind the load block here too. Ensure
that if the block return is taken, the ret variable
contains next_spec_k, so that processing continues with
whatever directive follows the @(load).

* txr.c (txr_main): Bind the block in severl cases that
load code.

* txr.1: Documented.