bug: autoload issue affecting with-compilation-unit
When a (with-compilation-unit ...) form is compiled,
it references the value of a certain special variable
that is private to the compiler. This reference is
correct, but the variable's symbol is not hooked up
to trigger autoload. If the compiler module has not
yet been loaded, and the compiled version of a
with-compilation-unit form is evaluated, it will fail
due to an unbound variable error.

* autoload.c (compiler_set_entries): Add the
sys:*in-compilation-unit* variable name to the autoload
triggers for the compiler module.