eval: take macro environment.
With this change we fix the bug that the debugger commands
yield their Lisp forms rather than evaluating them.

* eval.c (eval_intrinsic): Takes one more argument,
the macro environment. This is passed into env_to_menv
as the root macro environment.
(eval_init): Update registration of eval intrinsic
to have two optional arguments.

* eval.h (eval_intrinsic): Declaration updated.

* parser.c (read_file_common, read_eval_ret_last): Pass
nil argument to new parameter of eval_intrinsic.
(repl): Pass the env parameter as the new menv
parameter of eval_intrinsic, rather than the existing
env parameter. This fixes the command dispatch in
the debugger, since the command table is consists of
symbol macros, and not variables. For instance the
backtrace command bt is a binding of the bt symbol
to the form (sys:print-backtrace), which has to be
substituted for it and executed. When that envrionment
is used as the ordinary environment, bt looks like
a variable whose value is the list (sys:backtrace).

* parser.y (elem, check_parse_time_action): Fix
eval_intrinsic calls.

* txr.c (txr_main): Likewise.

* txr.1: Documented.

* y.tab.c.shipped: Updated.

* stdlib/doc-syms.tl: Updated.