New function: hash-map.
hash-map converts a function mapping over a sequence
into a hash table.

* hash.[ch] (hash_map): New function.

* tests/010/hash.tl: Test case.

* genman.txr: The hash-map identifier introduces
a hash collision. We have to deal with that somehow now.
(colli): We put the conflicting entries into a new hash called
colli which maps them to an increment value.
(hash-title): Increment the hash code h by the amount
indicated in colli, if the title is found there.

* txr.1: Documented.

* stdlib/doc-syms.tl: Updated.