Bug: ranges not treated as iterable in some situations.
* lib.c (seq_kind_tab): Map the RNG type to SEQ_VECLIKE
rather than SEQ_NOTSEQ. This one liner change causes
ranges to be rejected for iteration needlessly.
For instance, we can now do (take 42 "00".."99").
Note that ranges are not literally vector-like; they
don't support the ref operation. The effect of this change
is simply that they are not rejected due to being
SEQ_NOTSEQ. The iteration code already handles RNG
specially, so the SEQ_VECLIKE classification doesn't
actually matter.