crypt: handle libxcrypt failure tokens.
Some OS distros have switched to a library called libxcrypt,
which, instead of returning null on failure, like the Glibc
crypt, returns a valid "failure token" string starting with
a * character, which is guaranteed to be different from the
salt argument.

We should check for this so that we handle errors uniformly.
Users are reporting failing crypt tests that are coming
up with "*0" instead of throwing an exception.

* sysif.c (crypt_wrap): Only accept a non-null result if
it isn't one of the two strings "*0" and "*1".