bug: :vars not usable with :counter in @(repeat).
This is a regression due to a March 2016 commit which
introduced the ability for :vars in an output-side @(repeat)
block to have initial values.

The bug has the effect that all arguments in @(repeat)
which are conses/lists get duplicated, which messes up
the property list structure.

* parser.y (expand_repeat_rep_args): Do not unconditionally
add reg to the output at the bottom of the loop. A few
cases above in the consp(arg) case handle that themselves, and
do not continue the loop, so control ends up at the bottom,
adding a spurious item. By removing this list_collect,
we have to introduce it to just one case which relies on it.

* tests/008/repeat.txr,
* tests/008/repeat.expected: New files.

* y.tab.c.shipped: Updated.