New feature: local symbol renaming.
The new function use-sym-as can bring a foreign
symbol into a package under a different name,
which is not that symbol's name. This is also
featured in a new defpackage clause, :use-syms-as.

With this simple relaxation in the package system,
we don't require package local nicknames, which is
more complicated to implement and less ergonomic,
because it doesn't actually vanquish the use of
ugly package prefixes on clashing symbols.

* eval.c (eval_init): Register use-syms-as.

* lib.c (use_sym_as): New function, made out of
use_sym.
(use_sym): Now a wrapper for use_sym_as.

* lib.h (use_sym_as): Declared.

* stdlib/package.tl (defpackage): Implement :use-syms-as
clause.

* tests/012/use-as.tl: New file.

* txr.1: Documented,

* stdlib/doc-syms.tl: Updated.