build: remove mkdir commands from build steps.
* Makefile (COMPILE_C_WITH_DEPS, WINDRES): Do not call mkdir.
(win/%.res): Express dependency on existence of win
subdirectory in the build directory. We use an order-only
prerequisite, so that timestamps are not compared.
We don't want to be rebuilding objects if the directory
they are in has a newer timestamp, which is often the case.
(OBJS): Extract the relative directory paths from $(OBJS),
and make each directory the target of a rule which creates it.
Then, make each target in $(OBJS) dependent (order-only) on
the directory into which it will be placed, so before that
object is built, the directory gets created with mkdir -p.