configure: fix: _TIME_BITS is tied to _FILE_OFFSET_BITS
This problem was reported by Void Linux package maintainer
adigitoleo.

* configure: in glibc, _TIME_BITS and _FILE_OFFSET_BITS
are not independent. You cannot make the former 64 if
the latter isn't. I'm rewriting the test to check for
these together, in all combinations. We succeed if
we detect a combination that makes both time_t and
off_t 64 bits. If we don't find such a combination then
we at least enable 64 bit off_t alone or 64 bit time_t
alone, if we encountered an option which does that.