windows: don't require admin privilege.
* inst.nsi (RequestExecutionLevel): Use the value "highest"
instead of "admin", so that privilege is required for users
that have it, otherwise not.
(AccountType): New variable.
(.onInit): Obtain the account type and store it in AccountType.
Default the installation directory to $LOCALAPPDATA.
If the user is admin, change the default to the appropriate
Program Files directory.
(TXR): If the user isn't admin, add TXR to the current user's PATH
rather than the system PATH for everyone.
(Uninstall): Remove TXR from the appropriate PATH.