crypt: detect error tokens more weakly; drop some tests.
It has been reported by user cielesti that some of our
crypt tests fail on the Musl library.

Musl has some additional agorithms so it yields
a meaningful hash for a "$0$" salt, as well as for "$9$".

Musl uses "*" and "x" as error tokens rather than "*0"
and "*1". We need to change how we detect error tokens.

* sysif.c (crypt_wrap): Detect error tokens only by their
length: if a string emerges from crypt or crypt_r, whose
length is less than 13, it's an error token.

* tests/018/crypt.tl: Drop the tests that require :error
for salts "$0$" and "$9$", replacing them with a test
for a salt that is almost certainly invalid in all C libraries
on Linux.

* txr.1: Document that crypt throws an error exception and
under what circumstances (when the C library function does
what).